Wydzielenie klasy Product
