+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/OfferItem.java
@@ -20,92 +20,42 @@ import java.util.Date;
 
 public class OfferItem {
 
-	// product
-	private String productId;
-
-	private BigDecimal productPrice;
-
-	private String productName;
-
-	private Date productSnapshotDate;
-
-	private String productType;
-
-	private int quantity;
-
-	private BigDecimal totalCost;
-
-	private String currency;
+	private Product product;
 
 	// discount
 	private String discountCause;
 
 	private BigDecimal discount;
 
+	private String currency;
+
 	public OfferItem(String productId, BigDecimal productPrice, String productName,
 			Date productSnapshotDate, String productType, int quantity) {
 		this(productId, productPrice, productName, productSnapshotDate, productType, quantity, null, null);
 	}
 
-	public OfferItem(String productId, BigDecimal productPrice, String productName,
-			Date productSnapshotDate, String productType, int quantity,
-			BigDecimal discount, String discountCause) {
-		this.productId = productId;
-		this.productPrice = productPrice;
-		this.productName = productName;
-		this.productSnapshotDate = productSnapshotDate;
-		this.productType = productType;
+	public OfferItem(
+				String productId,
+				BigDecimal productPrice,
+				String productName,
+				Date productSnapshotDate,
+				String productType,
+				int quantity,
+				BigDecimal discount,
+				String discountCause
+			) {
 
-		this.quantity = quantity;
-		this.discount = discount;
-		this.discountCause = discountCause;
 
 		BigDecimal discountValue = new BigDecimal(0);
 		if (discount != null)
 			discountValue = discountValue.subtract(discount);
 
-		this.totalCost = productPrice
+		BigDecimal totalCost = productPrice
 				.multiply(new BigDecimal(quantity)).subtract(discountValue);
-	}
 
-	public String getProductId() {
-		return productId;
-	}
-	
-	public BigDecimal getProductPrice() {
-		return productPrice;
-	}
-	
-	public String getProductName() {
-		return productName;
-	}
-	
-	public Date getProductSnapshotDate() {
-		return productSnapshotDate;
-	}
-	
-	public String getProductType() {
-		return productType;
-	}
-
-	public BigDecimal getTotalCost() {
-		return totalCost;
-	}
-
-	public String getTotalCostCurrency() {
-		return currency;
-	}
-
-	public BigDecimal getDiscount() {
-		return discount;
-	}
-
-	public String getDiscountCause() {
-		return discountCause;
-	}
-
-	public int getQuantity() {
-		return quantity;
+		this.product = new Product(productId,productPrice,productName,productSnapshotDate,productType,quantity,totalCost);
+		this.discount = discount;
+		this.discountCause = discountCause;
 	}
 
 	@Override
