+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/Product.java
@@ -1,0 +1,90 @@
+package pl.com.bottega.ecommerce.sales.domain.offer;
+
+import java.math.BigDecimal;
+import java.util.Date;
+
+/**
+ * Created by Maciek on 07.03.2017.
+ */
+public class Product {
+
+    public Product(String productId, BigDecimal productPrice, String productName, Date productSnapshotDate, String productType, int quantity, BigDecimal totalCost) {
+        this.productId = productId;
+        this.productPrice = productPrice;
+        this.productName = productName;
+        this.productSnapshotDate = productSnapshotDate;
+        this.productType = productType;
+        this.quantity = quantity;
+        this.totalCost = totalCost;
+    }
+
+    private String productId;
+
+    private BigDecimal productPrice;
+
+    private String productName;
+
+    private Date productSnapshotDate;
+
+    private String productType;
+
+    private int quantity;
+
+    private BigDecimal totalCost;
+
+    public String getProductId() {
+        return productId;
+    }
+
+    public void setProductId(String productId) {
+        this.productId = productId;
+    }
+
+    public BigDecimal getProductPrice() {
+        return productPrice;
+    }
+
+    public void setProductPrice(BigDecimal productPrice) {
+        this.productPrice = productPrice;
+    }
+
+    public String getProductName() {
+        return productName;
+    }
+
+    public void setProductName(String productName) {
+        this.productName = productName;
+    }
+
+    public Date getProductSnapshotDate() {
+        return productSnapshotDate;
+    }
+
+    public void setProductSnapshotDate(Date productSnapshotDate) {
+        this.productSnapshotDate = productSnapshotDate;
+    }
+
+    public String getProductType() {
+        return productType;
+    }
+
+    public void setProductType(String productType) {
+        this.productType = productType;
+    }
+
+    public int getQuantity() {
+        return quantity;
+    }
+
+    public void setQuantity(int quantity) {
+        this.quantity = quantity;
+    }
+
+    public BigDecimal getTotalCost() {
+        return totalCost;
+    }
+
+    public void setTotalCost(BigDecimal totalCost) {
+        this.totalCost = totalCost;
+    }
+}
